test: split profile integration coverage
